+++ CHANGELOG.md
@@ -6,11 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
-## [2.0.0] - 2022-03-30
+## [2.0.0] - 2022-03-31
 
 ### Added
 
 - The new `SyntaxTree.register_handler` hook for plugins.
+- The new `--plugins=` option on the CLI.
 
 ### Changed